App/requirements.txt
langchain
langchain-community
openai
PyPDF2
faiss-cpu
tiktoken
streamlit
python-dotenv